✨[FEAT] alert 창 extension 빼주기
