+++ 29th-Assignment/29th-Assignment-Youtube/29th-Assignment-Youtube/Sources/Login/SignInVC.swift
@@ -85,32 +85,6 @@ extension SignInVC: UITextFieldDelegate {
 
 extension SignInVC {
     // Networking Alert
-    func successAlert(title: String, message: String) {
-        let alert = UIAlertController(title: title,
-                                      message: message,
-                                      preferredStyle: .alert)
-        let okAction = UIAlertAction(title: "확인", style: .default) { (action) in
-            guard let welcomeVC = self.storyboard?.instantiateViewController(withIdentifier: "WelcomeVC") as? WelcomeVC else {return}
-            
-            welcomeVC.userName = self.nameTextField.text
-            welcomeVC.modalPresentationStyle = .fullScreen
-            self.present(welcomeVC, animated: true, completion: nil)
-        }
-        
-        alert.addAction(okAction)
-        present(alert, animated: true)
-    }
-    
-    func failAlert(title: String, message: String) {
-        let alert = UIAlertController(title: title,
-                                      message: message,
-                                      preferredStyle: .alert)
-        let okAction = UIAlertAction(title: "확인", style: .default)
-        
-        alert.addAction(okAction)
-        present(alert, animated: true)
-    }
-    
     func requestLogin(){
         UserSignService.shared.login(email: emailTextField.text ?? "" , password: passwordTextField.text ?? "") { reponseData in
             switch reponseData {
