+++ 29th-Assignment/29th-Assignment-Youtube/29th-Assignment-Youtube/Sources/Login/SignUpVC.swift
@@ -92,33 +92,6 @@ extension SignUpVC: UITextFieldDelegate {
 }
 
 extension SignUpVC {
-    // Networking Alert
-    func successAlert(title: String, message: String) {
-        let alert = UIAlertController(title: title,
-                                      message: message,
-                                      preferredStyle: .alert)
-        let okAction = UIAlertAction(title: "확인", style: .default) { (action) in
-            guard let welcomeVC = self.storyboard?.instantiateViewController(withIdentifier: "WelcomeVC") as? WelcomeVC else {return}
-            
-            welcomeVC.userName = self.nameTextField.text
-            welcomeVC.modalPresentationStyle = .fullScreen
-            self.present(welcomeVC, animated: true, completion: nil)
-        }
-        
-        alert.addAction(okAction)
-        present(alert, animated: true)
-    }
-    
-    func failAlert(title: String, message: String) {
-        let alert = UIAlertController(title: title,
-                                      message: message,
-                                      preferredStyle: .alert)
-        let okAction = UIAlertAction(title: "확인", style: .default)
-        
-        alert.addAction(okAction)
-        present(alert, animated: true)
-    }
-    
     func requestSignUp(){
         UserSignService.shared.signUp(email: emailTextField.text ?? "", name: nameTextField.text ?? "" , password: passwordTextField.text ?? "") { reponseData in
             switch reponseData {
